Shift from terraform to cloudformation
