--- infrastructure/modules/ecr.tf
@@ -1,93 +1,0 @@
-resource "aws_ecr_repository" "dagit" {
-  name                 = "${var.platform_name}-dagit"
-  image_tag_mutability = "IMMUTABLE"
-}
-
-resource "aws_ecr_repository" "daemon" {
-  name                 = "${var.platform_name}-daemon"
-  image_tag_mutability = "IMMUTABLE"
-}
-
-resource "aws_ecr_repository" "user-code" {
-  name                 = "${var.platform_name}-user-code"
-  image_tag_mutability = "IMMUTABLE"
-}
-
-resource "aws_ecr_repository_policy" "dagit-repo-policy" {
-  repository = aws_ecr_repository.dagit.name
-  policy     = <<EOF
-  {
-    "Version": "2008-10-17",
-    "Statement": [
-      {
-        "Sid": "adds full ecr access to the demo repository",
-        "Effect": "Allow",
-        "Principal": "*",
-        "Action": [
-          "ecr:BatchCheckLayerAvailability",
-          "ecr:BatchGetImage",
-          "ecr:CompleteLayerUpload",
-          "ecr:GetDownloadUrlForLayer",
-          "ecr:GetLifecyclePolicy",
-          "ecr:InitiateLayerUpload",
-          "ecr:PutImage",
-          "ecr:UploadLayerPart"
-        ]
-      }
-    ]
-  }
-  EOF
-}
-
-resource "aws_ecr_repository_policy" "daemon-repo-policy" {
-  repository = aws_ecr_repository.daemon.name
-  policy     = <<EOF
-  {
-    "Version": "2008-10-17",
-    "Statement": [
-      {
-        "Sid": "adds full ecr access to the demo repository",
-        "Effect": "Allow",
-        "Principal": "*",
-        "Action": [
-          "ecr:BatchCheckLayerAvailability",
-          "ecr:BatchGetImage",
-          "ecr:CompleteLayerUpload",
-          "ecr:GetDownloadUrlForLayer",
-          "ecr:GetLifecyclePolicy",
-          "ecr:InitiateLayerUpload",
-          "ecr:PutImage",
-          "ecr:UploadLayerPart"
-        ]
-      }
-    ]
-  }
-  EOF
-}
-
-
-resource "aws_ecr_repository_policy" "user-code-repo-policy" {
-  repository = aws_ecr_repository.user-code.name
-  policy     = <<EOF
-  {
-    "Version": "2008-10-17",
-    "Statement": [
-      {
-        "Sid": "adds full ecr access to the demo repository",
-        "Effect": "Allow",
-        "Principal": "*",
-        "Action": [
-          "ecr:BatchCheckLayerAvailability",
-          "ecr:BatchGetImage",
-          "ecr:CompleteLayerUpload",
-          "ecr:GetDownloadUrlForLayer",
-          "ecr:GetLifecyclePolicy",
-          "ecr:InitiateLayerUpload",
-          "ecr:PutImage",
-          "ecr:UploadLayerPart"
-        ]
-      }
-    ]
-  }
-  EOF
-}
