--- infrastructure/modules/iam.tf
@@ -1,22 +1,0 @@
-resource "aws_iam_role" "ecs_task_execution_role" {
-  name = "${var.platform_name}_execution_role"
-
-  assume_role_policy = data.aws_iam_policy_document.assume_role_policy
-}
-
-data "aws_iam_policy_document" "assume_role_policy" {
-  statement {
-    actions = ["sts:AssumeRole"]
-
-    principals {
-      type        = "Service"
-      identifiers = ["ecs-tasks.amazonaws.com"]
-    }
-  }
-}
-
-resource "aws_iam_role" "task_role" {
-  name = "${var.platform_name}_task_role"
-
-  assume_role_policy = data.aws_iam_policy_document.assume_role_policy
-}
