--- infrastructure/modules/output.tf
@@ -1,3 +1,0 @@
-output "app_url" {
-  value = aws_alb.application_load_balancer.dns_name
-}
