--- infrastructure/modules/terraform.tfstate
@@ -1,184 +1,0 @@
-{
-  "version": 4,
-  "terraform_version": "1.5.0",
-  "serial": 7,
-  "lineage": "b7dc9f2b-63ef-1d76-740a-a1ac2100d6b2",
-  "outputs": {},
-  "resources": [
-    {
-      "mode": "managed",
-      "type": "aws_ecr_repository",
-      "name": "daemon",
-      "provider": "provider[\"registry.terraform.io/hashicorp/aws\"]",
-      "instances": [
-        {
-          "schema_version": 0,
-          "attributes": {
-            "arn": "arn:aws:ecr:eu-west-2:855530571579:repository/sfdata-platform-daemon",
-            "encryption_configuration": [
-              {
-                "encryption_type": "AES256",
-                "kms_key": ""
-              }
-            ],
-            "force_delete": null,
-            "id": "sfdata-platform-daemon",
-            "image_scanning_configuration": [
-              {
-                "scan_on_push": false
-              }
-            ],
-            "image_tag_mutability": "IMMUTABLE",
-            "name": "sfdata-platform-daemon",
-            "registry_id": "855530571579",
-            "repository_url": "855530571579.dkr.ecr.eu-west-2.amazonaws.com/sfdata-platform-daemon",
-            "tags": null,
-            "tags_all": {},
-            "timeouts": null
-          },
-          "sensitive_attributes": [],
-          "private": "eyJlMmJmYjczMC1lY2FhLTExZTYtOGY4OC0zNDM2M2JjN2M0YzAiOnsiZGVsZXRlIjoxMjAwMDAwMDAwMDAwfX0="
-        }
-      ]
-    },
-    {
-      "mode": "managed",
-      "type": "aws_ecr_repository",
-      "name": "dagit",
-      "provider": "provider[\"registry.terraform.io/hashicorp/aws\"]",
-      "instances": [
-        {
-          "schema_version": 0,
-          "attributes": {
-            "arn": "arn:aws:ecr:eu-west-2:855530571579:repository/sfdata-platform-dagit",
-            "encryption_configuration": [
-              {
-                "encryption_type": "AES256",
-                "kms_key": ""
-              }
-            ],
-            "force_delete": null,
-            "id": "sfdata-platform-dagit",
-            "image_scanning_configuration": [
-              {
-                "scan_on_push": false
-              }
-            ],
-            "image_tag_mutability": "IMMUTABLE",
-            "name": "sfdata-platform-dagit",
-            "registry_id": "855530571579",
-            "repository_url": "855530571579.dkr.ecr.eu-west-2.amazonaws.com/sfdata-platform-dagit",
-            "tags": null,
-            "tags_all": {},
-            "timeouts": null
-          },
-          "sensitive_attributes": [],
-          "private": "eyJlMmJmYjczMC1lY2FhLTExZTYtOGY4OC0zNDM2M2JjN2M0YzAiOnsiZGVsZXRlIjoxMjAwMDAwMDAwMDAwfX0="
-        }
-      ]
-    },
-    {
-      "mode": "managed",
-      "type": "aws_ecr_repository",
-      "name": "user-code",
-      "provider": "provider[\"registry.terraform.io/hashicorp/aws\"]",
-      "instances": [
-        {
-          "schema_version": 0,
-          "attributes": {
-            "arn": "arn:aws:ecr:eu-west-2:855530571579:repository/sfdata-platform-user-code",
-            "encryption_configuration": [
-              {
-                "encryption_type": "AES256",
-                "kms_key": ""
-              }
-            ],
-            "force_delete": null,
-            "id": "sfdata-platform-user-code",
-            "image_scanning_configuration": [
-              {
-                "scan_on_push": false
-              }
-            ],
-            "image_tag_mutability": "IMMUTABLE",
-            "name": "sfdata-platform-user-code",
-            "registry_id": "855530571579",
-            "repository_url": "855530571579.dkr.ecr.eu-west-2.amazonaws.com/sfdata-platform-user-code",
-            "tags": null,
-            "tags_all": {},
-            "timeouts": null
-          },
-          "sensitive_attributes": [],
-          "private": "eyJlMmJmYjczMC1lY2FhLTExZTYtOGY4OC0zNDM2M2JjN2M0YzAiOnsiZGVsZXRlIjoxMjAwMDAwMDAwMDAwfX0="
-        }
-      ]
-    },
-    {
-      "mode": "managed",
-      "type": "aws_ecr_repository_policy",
-      "name": "daemon-repo-policy",
-      "provider": "provider[\"registry.terraform.io/hashicorp/aws\"]",
-      "instances": [
-        {
-          "schema_version": 0,
-          "attributes": {
-            "id": "sfdata-platform-daemon",
-            "policy": "{\"Statement\":[{\"Action\":[\"ecr:BatchCheckLayerAvailability\",\"ecr:BatchGetImage\",\"ecr:CompleteLayerUpload\",\"ecr:GetDownloadUrlForLayer\",\"ecr:GetLifecyclePolicy\",\"ecr:InitiateLayerUpload\",\"ecr:PutImage\",\"ecr:UploadLayerPart\"],\"Effect\":\"Allow\",\"Principal\":\"*\",\"Sid\":\"adds full ecr access to the demo repository\"}],\"Version\":\"2008-10-17\"}",
-            "registry_id": "855530571579",
-            "repository": "sfdata-platform-daemon"
-          },
-          "sensitive_attributes": [],
-          "private": "bnVsbA==",
-          "dependencies": [
-            "aws_ecr_repository.daemon"
-          ]
-        }
-      ]
-    },
-    {
-      "mode": "managed",
-      "type": "aws_ecr_repository_policy",
-      "name": "dagit-repo-policy",
-      "provider": "provider[\"registry.terraform.io/hashicorp/aws\"]",
-      "instances": [
-        {
-          "schema_version": 0,
-          "attributes": {
-            "id": "sfdata-platform-dagit",
-            "policy": "{\"Statement\":[{\"Action\":[\"ecr:BatchCheckLayerAvailability\",\"ecr:BatchGetImage\",\"ecr:CompleteLayerUpload\",\"ecr:GetDownloadUrlForLayer\",\"ecr:GetLifecyclePolicy\",\"ecr:InitiateLayerUpload\",\"ecr:PutImage\",\"ecr:UploadLayerPart\"],\"Effect\":\"Allow\",\"Principal\":\"*\",\"Sid\":\"adds full ecr access to the demo repository\"}],\"Version\":\"2008-10-17\"}",
-            "registry_id": "855530571579",
-            "repository": "sfdata-platform-dagit"
-          },
-          "sensitive_attributes": [],
-          "private": "bnVsbA==",
-          "dependencies": [
-            "aws_ecr_repository.dagit"
-          ]
-        }
-      ]
-    },
-    {
-      "mode": "managed",
-      "type": "aws_ecr_repository_policy",
-      "name": "user-code-repo-policy",
-      "provider": "provider[\"registry.terraform.io/hashicorp/aws\"]",
-      "instances": [
-        {
-          "schema_version": 0,
-          "attributes": {
-            "id": "sfdata-platform-user-code",
-            "policy": "{\"Statement\":[{\"Action\":[\"ecr:BatchCheckLayerAvailability\",\"ecr:BatchGetImage\",\"ecr:CompleteLayerUpload\",\"ecr:GetDownloadUrlForLayer\",\"ecr:GetLifecyclePolicy\",\"ecr:InitiateLayerUpload\",\"ecr:PutImage\",\"ecr:UploadLayerPart\"],\"Effect\":\"Allow\",\"Principal\":\"*\",\"Sid\":\"adds full ecr access to the demo repository\"}],\"Version\":\"2008-10-17\"}",
-            "registry_id": "855530571579",
-            "repository": "sfdata-platform-user-code"
-          },
-          "sensitive_attributes": [],
-          "private": "bnVsbA==",
-          "dependencies": [
-            "aws_ecr_repository.user-code"
-          ]
-        }
-      ]
-    }
-  ],
-  "check_results": null
-}
